pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linguaedit"
version = "0.2.0"
description = "A GTK4 translation file editor for PO, TS, and JSON files"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    {name = "Daniel Nylander", email = "po@danielnylander.se"},
]
keywords = ["translation", "po", "ts", "json", "gettext", "i18n", "l10n", "gtk4"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Localization",
]
dependencies = [
    "PyGObject>=3.42",
    "polib>=1.2",
    "requests>=2.28",
    "pyenchant>=3.2",
]

[project.optional-dependencies]
ai = ["openai>=1.0", "anthropic>=0.18"]
dev = ["pytest", "ruff"]

[project.scripts]
linguaedit = "linguaedit.app:main"

[project.gui-scripts]
linguaedit-gui = "linguaedit.app:main"

[tool.setuptools.packages.find]
where = ["src"]
